+++ wgpu-core/src/command/mod.rs
@@ -83,6 +83,7 @@ pub(crate) use allocator::CommandAllocator;
 /// cbindgen:ignore
 pub use self::{compute_command::ComputeCommand, render_command::RenderCommand};
 
+
 pub(crate) use timestamp_writes::ArcPassTimestampWrites;
 pub use timestamp_writes::PassTimestampWrites;
 
@@ -804,6 +805,7 @@ pub(crate) struct BakedCommands {
     pub(crate) trackers: Tracker,
     pub(crate) temp_resources: Vec<TempResource>,
     pub(crate) indirect_draw_validation_resources: crate::indirect_validation::DrawResources,
+    pub(crate) multi_draw_resources: crate::multi_draw_emulation::MultiDrawResources,
     buffer_memory_init_actions: Vec<BufferInitTrackerAction>,
     texture_memory_actions: CommandBufferTextureMemoryActions,
 }
@@ -832,6 +834,7 @@ pub struct CommandBufferMutable {
     temp_resources: Vec<TempResource>,
 
     indirect_draw_validation_resources: crate::indirect_validation::DrawResources,
+    multi_draw_resources: crate::multi_draw_emulation::MultiDrawResources,
 
     pub(crate) commands: Vec<Command<ArcReferences>>,
 
@@ -848,6 +851,7 @@ impl CommandBufferMutable {
             trackers: self.trackers,
             temp_resources: self.temp_resources,
             indirect_draw_validation_resources: self.indirect_draw_validation_resources,
+            multi_draw_resources: self.multi_draw_resources,
             buffer_memory_init_actions: self.buffer_memory_init_actions,
             texture_memory_actions: self.texture_memory_actions,
         }
@@ -904,6 +908,9 @@ impl CommandEncoder {
                     temp_resources: Default::default(),
                     indirect_draw_validation_resources:
                         crate::indirect_validation::DrawResources::new(device.clone()),
+                    multi_draw_resources: crate::multi_draw_emulation::MultiDrawResources::new(
+                        device.clone(),
+                    ),
                     commands: Vec::new(),
                     #[cfg(feature = "trace")]
                     trace_commands: if device.trace.lock().is_some() {
@@ -1049,6 +1056,7 @@ impl CommandEncoder {
                     temp_resources: &mut cmd_buf_data.temp_resources,
                     indirect_draw_validation_resources: &mut cmd_buf_data
                         .indirect_draw_validation_resources,
+                    multi_draw_resources: &mut cmd_buf_data.multi_draw_resources,
                     snatch_guard: &snatch_guard,
                     debug_scope_depth: &mut debug_scope_depth,
                 };
@@ -1123,6 +1131,7 @@ impl CommandEncoder {
                     temp_resources: &mut cmd_buf_data.temp_resources,
                     indirect_draw_validation_resources: &mut cmd_buf_data
                         .indirect_draw_validation_resources,
+                    multi_draw_resources: &mut cmd_buf_data.multi_draw_resources,
                     snatch_guard: &snatch_guard,
                     debug_scope_depth: &mut debug_scope_depth,
                 };