+++ wgpu-types/src/instance.rs
@@ -203,10 +203,10 @@ bitflags::bitflags! {
         ///
         /// - When calling `dispatch_workgroups_indirect`, all 3 indirect arguments encoded in the buffer
         /// must be less than the `max_compute_workgroups_per_dimension` device limit.
-        /// - When calling `draw_indirect`/`draw_indexed_indirect`/`multi_draw_indirect`/`multi_draw_indexed_indirect`:
+        /// - When calling `draw_indirect`/`draw_indexed_indirect`/`multi_draw_indirect_count`/`multi_draw_indexed_indirect`:
         ///   - If `Features::INDIRECT_FIRST_INSTANCE` is not enabled on the device, the `first_instance` indirect argument must be 0.
         ///   - The `first_instance` & `instance_count` indirect arguments must form a range that fits within all bound vertex buffers with `step_mode` set to `Instance`.
-        /// - When calling `draw_indirect`/`multi_draw_indirect`:
+        /// - When calling `draw_indirect`/`multi_draw_indirect_count`:
         ///   - The `first_vertex` & `vertex_count` indirect arguments must form a range that fits within all bound vertex buffers with `step_mode` set to `Vertex`.
         /// - When calling `draw_indexed_indirect`/`multi_draw_indexed_indirect`:
         ///   - The `first_index` & `index_count` indirect arguments must form a range that fits within the bound index buffer.